requirements.txt
appnope==0.1.2
asttokens==2.0.5
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
backcall==0.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cryptography==36.0.1
cytoolz==0.11.2
debugpy==1.5.1
decorator==5.1.1
Deprecated==1.2.13
entrypoints==0.4
eth-abi==3.0.0
eth-hash==0.3.2
eth-typing==3.0.0
eth-utils==2.0.0
executing==0.8.3
Flask==2.0.3
h11==0.13.0
idna==3.3
injector==0.19.0
ipykernel==6.9.1
ipython==8.1.1
itsdangerous==2.1.0
jedi==0.18.1
Jinja2==3.0.3
jupyter-client==7.1.2
jupyter-core==4.9.2
lxml==4.8.0
MarkupSafe==2.1.0
matplotlib-inline==0.1.3
nest-asyncio==1.5.4
outcome==1.1.0
packaging==21.3
parsimonious==0.8.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.28
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.11.2
pyOpenSSL==22.0.0
pyparsing==3.0.8
PySocks==1.7.1
python-dateutil==2.8.2
PyYAML==6.0
pyzmq==22.3.0
redis==4.2.2
requests==2.27.1
selenium==4.1.2
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
stack-data==0.2.0
toml==0.10.2
toolz==0.11.2
tornado==6.1
traitlets==5.1.1
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.26.8
wcwidth==0.2.5
websocket-client==1.3.1
Werkzeug==2.0.3
wrapt==1.14.1
wsproto==1.1.0
